@@ -1,58 +1,26 @@
 {
   "name": "react-router-dom",
-  "version": "5.3.0",
-  "description": "DOM bindings for React Router",
-  "repository": "ReactTraining/react-router",
+  "version": "6.0.0",
+  "author": "Remix Software <hello@remix.run>",
+  "description": "Declarative routing for React web applications",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/remix-run/react-router.git",
+    "directory": "packages/react-router-dom"
+  },
   "license": "MIT",
-  "author": "React Training <hello@reacttraining.com>",
-  "files": [
-    "BrowserRouter.js",
-    "HashRouter.js",
-    "Link.js",
-    "MemoryRouter.js",
-    "NavLink.js",
-    "Prompt.js",
-    "Redirect.js",
-    "Route.js",
-    "Router.js",
-    "StaticRouter.js",
-    "Switch.js",
-    "cjs",
-    "es",
-    "esm",
-    "index.js",
-    "generatePath.js",
-    "matchPath.js",
-    "modules/*.js",
-    "modules/utils/*.js",
-    "withRouter.js",
-    "warnAboutDeprecatedCJSRequire.js",
-    "umd"
-  ],
-  "main": "index.js",
-  "module": "esm/react-router-dom.js",
-  "sideEffects": false,
-  "scripts": {
-    "build": "rollup -c",
-    "lint": "eslint modules"
+  "main": "./main.js",
+  "module": "./index.js",
+  "types": "./index.d.ts",
+  "unpkg": "./umd/react-router-dom.production.min.js",
+  "dependencies": {
+    "react-router": "6.0.0"
   },
   "peerDependencies": {
-    "react": ">=15"
+    "react": ">=16.8",
+    "react-dom": ">=16.8"
   },
-  "dependencies": {
-    "@babel/runtime": "^7.12.13",
-    "history": "^4.9.0",
-    "loose-envify": "^1.3.1",
-    "prop-types": "^15.6.2",
-    "react-router": "5.2.1",
-    "tiny-invariant": "^1.0.2",
-    "tiny-warning": "^1.0.0"
-  },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
-  },
+  "sideEffects": false,
   "keywords": [
     "react",
     "router",
