@@ -2,39 +2,43 @@ import React from "react";
 import "./Card.css";
 import { Link } from "react-router-dom";
 
-export default function Card({ min, max, name, img, onClose, id }) {
-  return (
-    <div className="card">
-      <div id="closeIcon" className="row">
-        <button onClick={onClose} className="btn btn-sm btn-danger">
-          X
-        </button>
-      </div>
+export default function Card({ min, max, name, img, onClose, id, getDetalle }) {
 
-      <div className="card-body">
-        <Link to={`ciudad/${id}`}>
-          <h5 className="card-title">{name}</h5>
-        </Link>
-        <div className="row">
-          <div className="col-sm-4 col-md-4 col-lg-4">
-            <p>Min</p>
-            <p>{min}°</p>
-          </div>
-          <div className="col-sm-4 col-md-4 col-lg-4">
-            <p>Max</p>
-            <p>{max}°</p>
-          </div>
-          <div className="col-sm-4 col-md-4 col-lg-4">
-            <img
-              className="iconoClima"
-              src={"http://openweathermap.org/img/wn/" + img + "@2x.png"}
-              width="80"
-              height="80"
-              alt=""
-            />
+  const handleOnClick = () =>{
+    getDetalle(id)
+  }
+
+  return (
+    <div>
+      {/* <Link to={`ciudad/${id}`}> */}
+        <div className="card_container" onClick={handleOnClick}>
+          <div className="card_header">
+            <div className="card_title">
+              <p>{name}</p>
+            </div>
+            <div id="closeIcon" >
+              <button onClick={onClose}>
+                X
+              </button>
+            </div>
+          </div>  
+          <div className="card_body">
+            <div>
+                <img
+                  className="iconoClima"
+                  src={"http://openweathermap.org/img/wn/" + img + "@2x.png"}
+                  width="70"
+                  height="70"
+                  alt=""
+                />
+            </div>
+            <div className="temperatura">
+                <p>Min: {min}°</p>
+                <p>Max: {max}°</p>
+            </div>
           </div>
         </div>
-      </div>
+      {/* </Link> */}
     </div>
   );
 }
