@@ -1,8 +1,6 @@
 import React from "react";
-import Logo from "../img/logoHenry.png";
 import SearchBar from "./SearchBar.jsx";
 import "./Nav.css";
-import About from "./About.jsx";
 import { Link } from "react-router-dom";
 
 function Nav({ onSearch }) {
@@ -10,21 +8,13 @@ function Nav({ onSearch }) {
     <nav className="navbar navbar-dark bg-dark">
       <Link to="/">
         <span className="navbar-brand">
-          <img
-            id="logoHenry"
-            src={Logo}
-            width="30"
-            height="30"
-            className="d-inline-block align-top"
-            alt=""
-          />
-          Henry - Weather App
+          Weather App
         </span>
       </Link>
       <Link to="/about">
         <span>About</span>
       </Link>
-      <SearchBar onSearch={onSearch} />
+      
     </nav>
   );
 }
