@@ -1,12 +1,16 @@
 import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
+
 
 export default function SearchBar({ onSearch }) {
   const [city, setCity] = useState("");
+  const navigate = useNavigate();
   return (
     <form
       onSubmit={(e) => {
         e.preventDefault();
         onSearch(city);
+        // navigate('/');
         setCity("");
       }}
     >
@@ -16,7 +20,9 @@ export default function SearchBar({ onSearch }) {
         value={city}
         onChange={(e) => setCity(e.target.value)}
       />
-      <input type="submit" value="Agregar" />
+      
+        <input type="submit" value="Agregar" />
+        
     </form>
   );
 }
